+++ Sources/Cell.swift
@@ -1,26 +1,8 @@
-//
-// Cell.swift
-//
-// Copyright (c) 2016 Marko Tadić <tadija@me.com> http://tadija.net
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-//
+/**
+ *  https://github.com/tadija/AEConsole
+ *  Copyright (c) Marko Tadić 2016-2018
+ *  Licensed under the MIT license. See LICENSE file.
+ */
 
 import UIKit
 
@@ -32,7 +14,7 @@ class Cell: UITableViewCell {
     
     // MARK: - Properties
     
-    private let config = Config.shared
+    private let settings = Console.shared.settings
     
     // MARK: - Init
     
@@ -54,8 +36,8 @@ class Cell: UITableViewCell {
     private func configureLabel() {
         guard let label = textLabel else { return }
         
-        label.font = config.consoleFont
-        label.textColor = config.textColorWithOpacity
+        label.font = settings.consoleFont
+        label.textColor = settings.textColorWithOpacity
         label.numberOfLines = 1
         label.textAlignment = .left
     }
@@ -65,7 +47,7 @@ class Cell: UITableViewCell {
     override func prepareForReuse() {
         super.prepareForReuse()
         
-        textLabel?.textColor = config.textColorWithOpacity
+        textLabel?.textColor = settings.textColorWithOpacity
     }
     
     override func layoutSubviews() {