+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Version 0.4.0
+
+- Migrated to Swift 4 and recommended settings with Xcode 9.2
+- Replaced `Config` (from .plist file) with `Settings` class (configure via code)
+- Renamed class `AEConsole` to `Console`
+
+## Version 0.3.1
+
+- Minor fixes
+
 ## Version 0.3.0
 
 - Fixed all build warnings and errors in Xcode 8.1 Beta (8T47)